forward pass for reservoir model

+++ src/kernels/mini_vandermonde.py
@@ -38,7 +38,7 @@ class MiniVandermonde(nn.Module):
         self.d_state = d_state
         self.d_output = self.d_input  # Necessary condition for the Vandermonde kernel (SISO)
 
-        self.register_buffer('x0', torch.zeros(self.d_state))
+        self.register_buffer('x0', torch.zeros(self.d_state, dtype=torch.complex64))
 
         input_output_reservoir = Reservoir(d_in=self.d_state, d_out=self.d_output)
 
@@ -130,6 +130,7 @@ class MiniVandermonde(nn.Module):
         :return: y: time step output of shape (B, H), x: time step state of shape (B, P)
         """
         with torch.no_grad():
+            u = u.to(dtype=torch.complex64)
             A = torch.view_as_complex(self.A)  # (P)
             W = torch.view_as_complex(self.W)  # (H, P)
             B = torch.sqrt(torch.transpose(W, 0, 1))  # (P, H)
@@ -299,7 +300,7 @@ class MiniVandermondeReservoir(nn.Module):
         self.d_state = d_state
         self.d_output = self.d_input  # Necessary condition for the Vandermonde kernel (SISO)
 
-        self.register_buffer('x0', torch.zeros(self.d_state))
+        self.register_buffer('x0', torch.zeros(self.d_state, dtype=torch.complex64))
 
         input_output_reservoir = Reservoir(d_in=self.d_state, d_out=self.d_output)
 
@@ -356,6 +357,7 @@ class MiniVandermondeReservoir(nn.Module):
         :return: y: time step output of shape (B, H), x: time step state of shape (B, P)
         """
         with torch.no_grad():
+            u = u.to(dtype=torch.complex64)
             if x is None:
                 x = self.x0.unsqueeze(0).expand(u.shape[0], -1)
             x = torch.einsum('p,bp->bp', self.A, x) + torch.einsum('ph,bh->bp', self.B, u)  # (B,P)
